\name{PGHorse-package}
\alias{PGHorse-package}
\alias{PGHorse}
\docType{package}
\title{
\packageTitle{PGHorse}
}
\description{
\packageDescription{PGHorse}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PGHorse}
\packageIndices{PGHorse}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PGHorse}

Maintainer: \packageMaintainer{PGHorse}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
