\name{PGHorse-package}
\alias{PGHorse-package}
\alias{PGHorse}
\docType{package}
\title{
\packageTitle{PGHorse}
}
\description{
\packageDescription{PGHorseS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PGHorse}
\packageIndices{PGHorse}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PGHorse}

Maintainer: \packageMaintainer{PGHorseS}
}
\references{~~ Literature or other references for background information ~~}
\keyword{ package }
\examples{~~ simple examples of the most important functions ~~}